ajustes no componente select-patient
