+++ src/components/side-menu/menu-itens/select-patient/index.js
@@ -1,12 +1,11 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import Box from "@mui/material/Box";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
 import Typography from "@mui/material/Typography";
 import { makeStyles } from "@mui/styles";
-import { fetchAll } from "../../../../services/api/patient";
 import { useMyContext } from "../../../../providers/MyContext";
 // import { ContactlessOutlined } from "@mui/icons-material";
 
@@ -32,31 +31,13 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const SelectPatient = () => {
+const SelectPatient = ({ patientList = []}) => {
   const context = useMyContext();
   const classes = useStyles();
-  const [patientId, setPatientId] = useState('');
-  const [patients, setPatients] = useState([])
-
-  useEffect(() => {
-    fetchAllPatients()
-  }, [])
-
-  const fetchAllPatients = async () => {
-    try {
-      const result = await fetchAll(context)
-      setPatients([...result]);
-      setPatientId(result[0].id || '');
-      context.setPatientName(result[0].name || '');
-      context.setPatientBirthDate(result[0].birthDate || '');
-      context.setPatientId(result[0].id || '');
-    } catch (error) { }
-  }
 
   const handlePatientChange = (event) => {
     const patientId = event.target.value;
-    setPatientId(patientId)
-    const patient = patients.find(elem => elem.id === patientId);
+    const patient = patientList.find(elem => elem.id === patientId);
     context.setPatientName(patient.name);
     context.setPatientBirthDate(patient.birthDate);
     context.setPatientId(patientId)
@@ -68,12 +49,12 @@ const SelectPatient = () => {
         <Select
           className={classes.select}
           size="small"
-          value={patientId}
+          value={context.patientId || ''}
           onChange={handlePatientChange}
           displayEmpty
           inputProps={{ "aria-label": "Without label" }}
         >
-          {patients.length && patients.map((patient) => (
+          {patientList.length && patientList.map((patient) => (
             <MenuItem
               key={`SelectPatient${patient.id}`}
               value={patient.id}
