+++ src/pages/layout/index.js
@@ -1,5 +1,6 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 // import React from "react";
-import * as React from "react";
+import React, { useState, useEffect } from "react";
 import Constants from "../../utils/contants";
 import { Outlet } from "react-router-dom";
 import Box from "@mui/material/Box";
@@ -9,8 +10,36 @@ import TemporaryDrawer from "../../components/side-menu/temporary-drawer";
 import PermanentDrawer from "../../components/side-menu/permanent-drawer";
 import MenuItens from "../../components/side-menu/menu-itens";
 
+import { getTokenParameters } from '../../providers/sessionStorage';
+import { useMyContext } from "../../providers/MyContext";
+import { fetchAll } from "../../services/api/patient";
+
+
 function LayoutPage(props) {
-  const [mobileOpen, setMobileOpen] = React.useState(false);
+  const context = useMyContext();
+  const [permission] = useState(getTokenParameters('role') === "Administrator");
+  const [mobileOpen, setMobileOpen] = useState(false);
+  const [patientList, setPatientList] = useState([]);
+ 
+  const setPatientAccount = async () => {
+    context.setPatientId(getTokenParameters('pacientId'));
+    context.setPatientName(getTokenParameters('fullname'));
+  }
+
+  const fetchAllPatients = async () => {
+    try {
+      const result = await fetchAll(context)
+      setPatientList([...result]);
+      context.setPatientName(result[0].name || '');
+      context.setPatientBirthDate(result[0].birthDate || '');
+      context.setPatientId(result[0].id || '');
+    } catch (error) { }
+  }
+
+  useEffect(() => {
+    if (!permission) setPatientAccount();
+    else fetchAllPatients();
+  }, []);
 
   const SideBar = (
     <>
@@ -24,11 +53,11 @@ function LayoutPage(props) {
           open={mobileOpen}
           onCloseCallback={() => setMobileOpen(!mobileOpen)}
         >
-          <MenuItens />
+          <MenuItens permission={permission} patientList={patientList} />
         </TemporaryDrawer>
-        <PermanentDrawer width={Constants.DRAWER_WIDTH}>
-          <MenuItens />
-        </PermanentDrawer>
+        <PermanentDrawer width={Constants.DRAWER_WIDTH} >
+          <MenuItens permission={permission} patientList={patientList} />
+        </PermanentDrawer> 
       </Box>
     </>
   );
